fix: footer css
