+++ src/component/Footer.jsx
@@ -6,18 +6,17 @@ import Group37 from "../assets/Footer/Group 37.svg"
 import Group38 from "../assets/Footer/Group 38.svg"
 import Group39 from "../assets/Footer/Group 39.svg"
 
-
-export default function Footer() {
+export default function Footerm() {
     const imgsSrc = [Group33, Group34, Group35, Group36, Group37, Group38, Group39]
+
     return (
-        <div className="flex bg-green px-12 py-7 items-center">
-            <div className="flex space-x-2 mr-auto">
-                {imgsSrc.map(src =>
-                    (<img className="w-3.5 h-3.5" src />)
+        <footer className="flex-col items-center justify-center space-y-2 mt-auto py-3  bg-green sm:flex-row sm:px-20 sm:py-10" >
+            <div className="flex items-center justify-center space-x-2 sm:mr-auto">
+                {imgsSrc.map((imgSrc, index) =>
+                    (<img key={index} className="w-3.5 h-3.5" src={imgSrc} />)
                 )}
             </div>
-
-            <p class="text-white text-xs">© Start, 2022. All rights reserved.</p>
-        </div>
+            <p className="text-center text-white text-xs">© Start, 2022. All rights reserved.</p>
+        </footer>
     )
 }
